requirements.txt
# FastMCP Gmail Server Dependencies
# Python 3.10+ required

# FastMCP framework for MCP server
fastmcp>=2.13.0

# Google API libraries
google-api-python-client>=2.100.0
google-auth>=2.25.0

# Environment variable management
python-dotenv>=1.0.0
